Implement site generation
To make this simpler, asset files are combined with source files,
within the same directory. These files are then copied to the output dir.

The only different is markdown files are parsed using Discount, and since
we're not using a pre-processor, CSS files are just copied from memory.

This change required me to change the default route for css to be local.